+++ features/common.py
@@ -1,0 +1,7 @@
+from behave import *
+
+
+@when('we input command "{command}"')
+def step_impl(context, command):
+    command = command.split(' ')
+    context.command += [command[0], command[1]]