+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
+Release 0.1.0-dev
+---------------------------------
+* #1 Report generation about test runs in MD format (@flipback)
+* #7 A user can comment the failed step (@flipback)
+* #5 A user can edit a new test case in editor at once after creation (@flipback)
+
 Release 0.0.3
 ---------------------------------
 
 * Fix input on Windows (@flipback)
 * Fix saving a report at the end of command *add* (@flipback)
-